Swap Base64 implementation with Apache Commons Codec
ref #8
