--- modules/bcrypt/src/main/java/at/favre/lib/crypto/bcrypt/Radix64ApacheCodec.java
@@ -1,568 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package at.favre.lib.crypto.bcrypt;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Arrays;
-
-/**
- * Provides Radix64ApacheCodec encoding and decoding as defined by <a href="http://www.ietf.org/rfc/rfc2045.txt">RFC 2045</a>.
- *
- * <p>
- * This class implements section <cite>6.8. Radix64ApacheCodec Content-Transfer-Encoding</cite> from RFC 2045 <cite>Multipurpose
- * Internet Mail Extensions (MIME) Part One: Format of Internet Message Bodies</cite> by Freed and Borenstein.
- * </p>
- * <p>
- * The class can be parameterized in the following manner with various constructors:
- * </p>
- * <ul>
- * <li>URL-safe mode: Default off.</li>
- * <li>Line length: Default 76. Line length that aren't multiples of 4 will still essentially end up being multiples of
- * 4 in the encoded data.
- * <li>Line separator: Default is CRLF ("\r\n")</li>
- * </ul>
- * <p>
- * The URL-safe parameter is only applied to encode operations. Decoding seamlessly handles both modes.
- * </p>
- * <p>
- * Since this class operates directly on byte streams, and not character streams, it is hard-coded to only
- * encode/decode character encodings which are compatible with the lower 127 ASCII chart (ISO-8859-1, Windows-1252,
- * UTF-8, etc).
- * </p>
- * <p>
- * This class is thread-safe.
- * </p>
- *
- * @version $Id: Radix64ApacheCodec.java 1789158 2017-03-28 15:04:58Z sebb $
- * @see <a href="http://www.ietf.org/rfc/rfc2045.txt">RFC 2045</a>
- * @since 1.0
- */
-public class Radix64ApacheCodec implements Radix64Encoder {
-
-
-    /**
-     * Chunk separator per RFC 2045 section 2.1.
-     *
-     * <p>
-     * N.B. The next major release may break compatibility and make this field private.
-     * </p>
-     *
-     * @see <a href="http://www.ietf.org/rfc/rfc2045.txt">RFC 2045 section 2.1</a>
-     */
-    private static final byte[] CHUNK_SEPARATOR = {'\r', '\n'};
-    /**
-     * BASE32 characters are 6 bits in length.
-     * They are formed by taking a block of 3 octets to form a 24-bit string,
-     * which is converted into 4 BASE64 characters.
-     */
-    private static final int BITS_PER_ENCODED_BYTE = 6;
-    private static final int BYTES_PER_UNENCODED_BLOCK = 3;
-    private static final int BYTES_PER_ENCODED_BLOCK = 4;
-    /**
-     * This array is a lookup table that translates 6-bit positive integer index values into their "Radix64ApacheCodec Alphabet"
-     * equivalents as specified in Table 1 of RFC 2045.
-     * <p>
-     * Thanks to "commons" project in ws.apache.org for this code.
-     * http://svn.apache.org/repos/asf/webservices/commons/trunk/modules/util/
-     */
-    private static final byte[] STANDARD_ENCODE_TABLE = {
-            '.', '/', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J',
-            'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V',
-            'W', 'X', 'Y', 'Z', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h',
-            'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't',
-            'u', 'v', 'w', 'x', 'y', 'z', '0', '1', '2', '3', '4', '5',
-            '6', '7', '8', '9'
-    };
-
-    /**
-     * This array is a lookup table that translates Unicode characters drawn from the "Radix64ApacheCodec Alphabet" (as specified
-     * in Table 1 of RFC 2045) into their 6-bit positive integer equivalents. Characters that are not in the Radix64ApacheCodec
-     * alphabet but fall within the bounds of the array are translated to -1.
-     * <p>
-     * Note: '+' and '-' both decode to 62. '/' and '_' both decode to 63. This means decoder seamlessly handles both
-     * URL_SAFE and STANDARD base64. (The encoder, on the other hand, needs to know ahead of time what to emit).
-     * <p>
-     * Thanks to "commons" project in ws.apache.org for this code.
-     * http://svn.apache.org/repos/asf/webservices/commons/trunk/modules/util/
-     */
-    private static final byte[] DECODE_TABLE = {
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, 0, 1, 54, 55, 56, 57,
-            58, 59, 60, 61, 62, 63, -1, -1, -1, -2, -1, -1, -1, 2, 3, 4, 5, 6, 7,
-            8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25,
-            26, 27, -1, -1, -1, -1, -1, -1, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37,
-            38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1,
-            -1, -1, -1, -1, -1
-    };
-
-    /**
-     * Mask used to extract 6 bits, used when encoding
-     */
-    private static final int MASK_6BITS = 0x3f;
-    private static final int DEFAULT_BUFFER_RESIZE_FACTOR = 2;
-    /**
-     * Defines the default buffer size - currently {@value}
-     * - must be large enough for at least one encoded block+separator
-     */
-    private static final int DEFAULT_BUFFER_SIZE = 8192;
-    /**
-     * Mask used to extract 8 bits, used in decoding bytes
-     */
-    private static final int MASK_8BITS = 0xff;
-    /**
-     * Encode table to use: either STANDARD or URL_SAFE. Note: the DECODE_TABLE above remains static because it is able
-     * to decode both STANDARD and URL_SAFE streams, but the encodeTable must be a member variable so we can switch
-     * between the two modes.
-     */
-    private final byte[] encodeTable;
-    /**
-     * Line separator for encoding. Not used when decoding. Only used if lineLength &gt; 0.
-     */
-    private final byte[] lineSeparator;
-    /**
-     * Convenience variable to help us determine when our buffer is going to run out of room and needs resizing.
-     * <code>decodeSize = 3 + lineSeparator.length;</code>
-     */
-    private final int decodeSize;
-    /**
-     * Convenience variable to help us determine when our buffer is going to run out of room and needs resizing.
-     * <code>encodeSize = 4 + lineSeparator.length;</code>
-     */
-    private final int encodeSize;
-    private final byte pad; // instance variable just in case it needs to vary later
-
-    /**
-     * Chunksize for encoding. Not used when decoding.
-     * A value of zero or less implies no chunking of the encoded data.
-     * Rounded down to nearest multiple of encodedBlockSize.
-     */
-    private final int lineLength;
-
-    /**
-     * Creates a Radix64ApacheCodec codec used for decoding (all modes) and encoding in URL-unsafe mode.
-     * <p>
-     * When encoding the line length is 0 (no chunking), and the encoding table is STANDARD_ENCODE_TABLE.
-     * </p>
-     *
-     * <p>
-     * When decoding all variants are supported.
-     * </p>
-     */
-    Radix64ApacheCodec() {
-        this(0, CHUNK_SEPARATOR);
-    }
-
-    /**
-     * Creates a Radix64ApacheCodec codec used for decoding (all modes) and encoding in URL-unsafe mode.
-     * <p>
-     * When encoding the line length and line separator are given in the constructor, and the encoding table is
-     * STANDARD_ENCODE_TABLE.
-     * </p>
-     * <p>
-     * Line lengths that aren't multiples of 4 will still essentially end up being multiples of 4 in the encoded data.
-     * </p>
-     * <p>
-     * When decoding all variants are supported.
-     * </p>
-     *
-     * @param lineLength    Each line of encoded data will be at most of the given length (rounded down to nearest multiple of
-     *                      4). If lineLength &lt;= 0, then the output will not be divided into lines (chunks). Ignored when
-     *                      decoding.
-     * @param lineSeparator Each line of encoded data will end with this sequence of bytes.
-     * @throws IllegalArgumentException The provided lineSeparator included some base64 characters. That's not going to work!
-     * @since 1.4
-     */
-    private Radix64ApacheCodec(final int lineLength, final byte[] lineSeparator) {
-        int encodedBlockSize = BYTES_PER_ENCODED_BLOCK;
-        int chunkSeparatorLength = lineSeparator == null ? 0 : lineSeparator.length;
-        final boolean useChunking = lineLength > 0 && chunkSeparatorLength > 0;
-        this.lineLength = useChunking ? (lineLength / encodedBlockSize) * encodedBlockSize : 0;
-        this.pad = '=';
-
-        if (lineSeparator != null) {
-            if (containsAlphabetOrPad(lineSeparator)) {
-                final String sep = newStringUtf8(lineSeparator);
-                throw new IllegalArgumentException("lineSeparator must not contain base64 characters: [" + sep + "]");
-            }
-            if (lineLength > 0) { // null line-sep forces no chunking rather than throwing IAE
-                this.encodeSize = BYTES_PER_ENCODED_BLOCK + lineSeparator.length;
-                this.lineSeparator = new byte[lineSeparator.length];
-                System.arraycopy(lineSeparator, 0, this.lineSeparator, 0, lineSeparator.length);
-            } else {
-                this.encodeSize = BYTES_PER_ENCODED_BLOCK;
-                this.lineSeparator = null;
-            }
-        } else {
-            this.encodeSize = BYTES_PER_ENCODED_BLOCK;
-            this.lineSeparator = null;
-        }
-        this.decodeSize = this.encodeSize - 1;
-        this.encodeTable = STANDARD_ENCODE_TABLE;
-    }
-
-    private static String newStringUtf8(final byte[] bytes) {
-        return bytes == null ? null : new String(bytes, StandardCharsets.UTF_8);
-    }
-
-    /**
-     * Tests a given byte array to see if it contains any characters within the alphabet or PAD.
-     * <p>
-     * Intended for use in checking line-ending arrays
-     *
-     * @param arrayOctet byte array to test
-     * @return <code>true</code> if any byte is a valid character in the alphabet or PAD; <code>false</code> otherwise
-     */
-    private boolean containsAlphabetOrPad(final byte[] arrayOctet) {
-        if (arrayOctet == null) {
-            return false;
-        }
-        for (final byte element : arrayOctet) {
-            if (pad == element || isInAlphabet(element)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Returns whether or not the <code>octet</code> is in the Radix64ApacheCodec alphabet.
-     *
-     * @param octet The value to test
-     * @return <code>true</code> if the value is defined in the the Radix64ApacheCodec alphabet <code>false</code> otherwise.
-     */
-    private boolean isInAlphabet(final byte octet) {
-        return octet >= 0 && DECODE_TABLE[octet] != -1;
-    }
-
-    /**
-     * <p>
-     * Encodes all of the provided data, starting at inPos, for inAvail bytes. Must be called at least twice: once with
-     * the data to encode, and once with inAvail set to "-1" to alert encoder that EOF has been reached, to flush last
-     * remaining bytes (if not multiple of 3).
-     * </p>
-     * <p><b>Note: no padding is added when encoding using the URL-safe alphabet.</b></p>
-     * <p>
-     * Thanks to "commons" project in ws.apache.org for the bitwise operations, and general approach.
-     * http://svn.apache.org/repos/asf/webservices/commons/trunk/modules/util/
-     * </p>
-     *
-     * @param in      byte[] array of binary data to base64 encode.
-     * @param inPos   Position to start reading data from.
-     * @param inAvail Amount of bytes available from input for encoding.
-     * @param context the context to be used
-     */
-    private void encode(final byte[] in, int inPos, final int inAvail, final Context context) {
-        if (context.eof) {
-            return;
-        }
-        // inAvail < 0 is how we're informed of EOF in the underlying data we're
-        // encoding.
-        if (inAvail < 0) {
-            context.eof = true;
-            if (0 == context.modulus && lineLength == 0) {
-                return; // no leftovers to process and not using chunking
-            }
-            final byte[] buffer = ensureBufferSize(encodeSize, context);
-            final int savedPos = context.pos;
-            switch (context.modulus) { // 0-2
-                case 0: // nothing to do here
-                    break;
-                case 1: // 8 bits = 6 + 2
-                    // top 6 bits:
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 2) & MASK_6BITS];
-                    // remaining 2:
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea << 4) & MASK_6BITS];
-                    break;
-
-                case 2: // 16 bits = 6 + 6 + 4
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 10) & MASK_6BITS];
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 4) & MASK_6BITS];
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea << 2) & MASK_6BITS];
-                    break;
-                default:
-                    throw new IllegalStateException("Impossible modulus " + context.modulus);
-            }
-            context.currentLinePos += context.pos - savedPos; // keep track of current line position
-            // if currentPos == 0 we are at the start of a line, so don't add CRLF
-            if (lineLength > 0 && context.currentLinePos > 0) {
-                System.arraycopy(lineSeparator, 0, buffer, context.pos, lineSeparator.length);
-                context.pos += lineSeparator.length;
-            }
-        } else {
-            for (int i = 0; i < inAvail; i++) {
-                final byte[] buffer = ensureBufferSize(encodeSize, context);
-                context.modulus = (context.modulus + 1) % BYTES_PER_UNENCODED_BLOCK;
-                int b = in[inPos++];
-                if (b < 0) {
-                    b += 256;
-                }
-                context.ibitWorkArea = (context.ibitWorkArea << 8) + b; //  BITS_PER_BYTE
-                if (0 == context.modulus) { // 3 bytes = 24 bits = 4 * 6 bits to extract
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 18) & MASK_6BITS];
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 12) & MASK_6BITS];
-                    buffer[context.pos++] = encodeTable[(context.ibitWorkArea >> 6) & MASK_6BITS];
-                    buffer[context.pos++] = encodeTable[context.ibitWorkArea & MASK_6BITS];
-                    context.currentLinePos += BYTES_PER_ENCODED_BLOCK;
-                    if (lineLength > 0 && lineLength <= context.currentLinePos) {
-                        System.arraycopy(lineSeparator, 0, buffer, context.pos, lineSeparator.length);
-                        context.pos += lineSeparator.length;
-                        context.currentLinePos = 0;
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Ensure that the buffer has room for <code>size</code> bytes
-     *
-     * @param size    minimum spare space required
-     * @param context the context to be used
-     * @return the buffer
-     */
-    private byte[] ensureBufferSize(final int size, final Context context) {
-        if ((context.buffer == null) || (context.buffer.length < context.pos + size)) {
-            return resizeBuffer(context);
-        }
-        return context.buffer;
-    }
-
-    /**
-     * Increases our buffer by the {@link #DEFAULT_BUFFER_RESIZE_FACTOR}.
-     *
-     * @param context the context to be used
-     */
-    private byte[] resizeBuffer(final Context context) {
-        if (context.buffer == null) {
-            context.buffer = new byte[DEFAULT_BUFFER_SIZE];
-            context.pos = 0;
-            context.readPos = 0;
-        } else {
-            final byte[] b = new byte[context.buffer.length * DEFAULT_BUFFER_RESIZE_FACTOR];
-            System.arraycopy(context.buffer, 0, b, 0, context.buffer.length);
-            context.buffer = b;
-        }
-        return context.buffer;
-    }
-
-    /**
-     * <p>
-     * Decodes all of the provided data, starting at inPos, for inAvail bytes. Should be called at least twice: once
-     * with the data to decode, and once with inAvail set to "-1" to alert decoder that EOF has been reached. The "-1"
-     * call is not necessary when decoding, but it doesn't hurt, either.
-     * </p>
-     * <p>
-     * Ignores all non-base64 characters. This is how chunked (e.g. 76 character) data is handled, since CR and LF are
-     * silently ignored, but has implications for other bytes, too. This method subscribes to the garbage-in,
-     * garbage-out philosophy: it will not check the provided data for validity.
-     * </p>
-     * <p>
-     * Thanks to "commons" project in ws.apache.org for the bitwise operations, and general approach.
-     * http://svn.apache.org/repos/asf/webservices/commons/trunk/modules/util/
-     * </p>
-     *
-     * @param in      byte[] array of ascii data to base64 decode.
-     * @param inPos   Position to start reading data from.
-     * @param inAvail Amount of bytes available from input for encoding.
-     * @param context the context to be used
-     */
-    private void decode(final byte[] in, int inPos, final int inAvail, final Context context) {
-        if (context.eof) {
-            return;
-        }
-        if (inAvail < 0) {
-            context.eof = true;
-        }
-        for (int i = 0; i < inAvail; i++) {
-            final byte[] buffer = ensureBufferSize(decodeSize, context);
-            final byte b = in[inPos++];
-            if (b == pad) {
-                // We're done.
-                context.eof = true;
-                break;
-            }
-            if (b >= 0) {
-                final int result = DECODE_TABLE[b];
-                if (result >= 0) {
-                    context.modulus = (context.modulus + 1) % BYTES_PER_ENCODED_BLOCK;
-                    context.ibitWorkArea = (context.ibitWorkArea << BITS_PER_ENCODED_BYTE) + result;
-                    if (context.modulus == 0) {
-                        buffer[context.pos++] = (byte) ((context.ibitWorkArea >> 16) & MASK_8BITS);
-                        buffer[context.pos++] = (byte) ((context.ibitWorkArea >> 8) & MASK_8BITS);
-                        buffer[context.pos++] = (byte) (context.ibitWorkArea & MASK_8BITS);
-                    }
-                }
-            }
-        }
-
-        // Two forms of EOF as far as base64 decoder is concerned: actual
-        // EOF (-1) and first time '=' character is encountered in stream.
-        // This approach makes the '=' padding characters completely optional.
-        if (context.eof && context.modulus != 0) {
-            final byte[] buffer = ensureBufferSize(decodeSize, context);
-
-            // We have some spare bits remaining
-            // Output all whole multiples of 8 bits and ignore the rest
-            switch (context.modulus) {
-                // case 0 : // impossible, as excluded above
-                case 1: // 6 bits - ignore entirely
-                    // TODO not currently tested; perhaps it is impossible?
-                    break;
-                case 2: // 12 bits = 8 + 4
-                    context.ibitWorkArea = context.ibitWorkArea >> 4; // dump the extra 4 bits
-                    buffer[context.pos++] = (byte) ((context.ibitWorkArea) & MASK_8BITS);
-                    break;
-                case 3: // 18 bits = 8 + 8 + 2
-                    context.ibitWorkArea = context.ibitWorkArea >> 2; // dump 2 bits
-                    buffer[context.pos++] = (byte) ((context.ibitWorkArea >> 8) & MASK_8BITS);
-                    buffer[context.pos++] = (byte) ((context.ibitWorkArea) & MASK_8BITS);
-                    break;
-                default:
-                    throw new IllegalStateException("Impossible modulus " + context.modulus);
-            }
-        }
-    }
-
-    @Override
-    public byte[] encode(byte[] rawBytes) {
-        final Context c = new Context();
-        encode(rawBytes, 0, rawBytes.length, c);
-        encode(rawBytes, 0, -1, c); // Notify encoder of EOF.
-        final byte[] buf = new byte[c.pos - c.readPos];
-        readResults(buf, 0, buf.length, c);
-        return buf;
-    }
-
-    /**
-     * Extracts buffered data into the provided byte[] array, starting at position bPos, up to a maximum of bAvail
-     * bytes. Returns how many bytes were actually extracted.
-     * <p>
-     * Package protected for access from I/O streams.
-     *
-     * @param b       byte[] array to extract the buffered data into.
-     * @param bPos    position in byte[] array to start extraction at.
-     * @param bAvail  amount of bytes we're allowed to extract. We may extract fewer (if fewer are available).
-     * @param context the context to be used
-     * @return The number of bytes successfully extracted into the provided byte[] array.
-     */
-    private int readResults(final byte[] b, final int bPos, final int bAvail, final Context context) {
-        if (context.buffer != null) {
-            final int len = Math.min(available(context), bAvail);
-            System.arraycopy(context.buffer, context.readPos, b, bPos, len);
-            context.readPos += len;
-            if (context.readPos >= context.pos) {
-                context.buffer = null; // so hasData() will return false, and this method can return -1
-            }
-            return len;
-        }
-        return context.eof ? -1 : 0;
-    }
-
-    private int available(final Context context) {  // package protected for access from I/O streams
-        return context.buffer != null ? context.pos - context.readPos : 0;
-    }
-
-    @Override
-    public byte[] decode(byte[] utf8EncodedRadix64String) {
-        final Context c = new Context();
-        decode(utf8EncodedRadix64String, 0, utf8EncodedRadix64String.length, c);
-        decode(utf8EncodedRadix64String, 0, -1, c); // Notify decoder of EOF.
-        final byte[] result = new byte[c.pos];
-        readResults(result, 0, result.length, c);
-        return result;
-    }
-
-    /**
-     * Holds thread context so classes can be thread-safe.
-     * <p>
-     * This class is not itself thread-safe; each thread must allocate its own copy.
-     *
-     * @since 1.7
-     */
-    static class Context {
-
-        /**
-         * Place holder for the bytes we're dealing with for our based logic.
-         * Bitwise operations store and extract the encoding or decoding from this variable.
-         */
-        int ibitWorkArea;
-
-        /**
-         * Place holder for the bytes we're dealing with for our based logic.
-         * Bitwise operations store and extract the encoding or decoding from this variable.
-         */
-        long lbitWorkArea;
-
-        /**
-         * Buffer for streaming.
-         */
-        byte[] buffer;
-
-        /**
-         * Position where next character should be written in the buffer.
-         */
-        int pos;
-
-        /**
-         * Position where next character should be read from the buffer.
-         */
-        int readPos;
-
-        /**
-         * Boolean flag to indicate the EOF has been reached. Once EOF has been reached, this object becomes useless,
-         * and must be thrown away.
-         */
-        boolean eof;
-
-        /**
-         * Variable tracks how many characters have been written to the current line. Only used when encoding. We use
-         * it to make sure each encoded line never goes beyond lineLength (if lineLength &gt; 0).
-         */
-        int currentLinePos;
-
-        /**
-         * Writes to the buffer only occur after every 3/5 reads when encoding, and every 4/8 reads when decoding. This
-         * variable helps track that.
-         */
-        int modulus;
-
-        Context() {
-        }
-
-        /**
-         * Returns a String useful for debugging (especially within a debugger.)
-         *
-         * @return a String useful for debugging.
-         */
-        @SuppressWarnings("boxing") // OK to ignore boxing here
-        @Override
-        public String toString() {
-            return String.format("%s[buffer=%s, currentLinePos=%s, eof=%s, ibitWorkArea=%s, lbitWorkArea=%s, " +
-                            "modulus=%s, pos=%s, readPos=%s]", this.getClass().getSimpleName(), Arrays.toString(buffer),
-                    currentLinePos, eof, ibitWorkArea, lbitWorkArea, modulus, pos, readPos);
-        }
-    }
-}
